# backend-node/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 宽窄·Orbit 业务后端 API
  description: |
    宽窄·Orbit 社区型 APP 业务后端 API 契约。

    ## 认证
    使用 Bearer Token 方式：`Authorization: Bearer <access_token>`

    ## 统一响应格式
    ```json
    { "success": true, "data": { ... } }
    ```
    错误时：
    ```json
    { "success": false, "error": { "code": "40001", "message": "..." }, "requestId": "req_xxx" }
    ```

    ## 错误码
    | Code | 含义 |
    |------|------|
    | 40001 | 请求参数错误 |
    | 40101 | 未登录或 Token 失效 |
    | 40301 | 无权限或被限制 |
    | 40401 | 资源不存在 |
    | 40901 | 状态冲突 / 重复操作 / 额度不足 |
    | 42901 | 请求频率过高 |
    | 50000 | 服务内部错误 |
    | 50401 | 上游超时 |

    ## 幂等
    写操作推荐携带 `Idempotency-Key` 请求头（UUID），相同 Key 的重复请求返回相同结果。

    ## 分页
    使用 page-based 分页：`?page=1&pageSize=20`，响应包含 `hasMore` 和 `nextPage`。
  version: 1.0.0
  contact:
    name: 宽窄·Orbit Backend Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.kuanzhai-orbit.com/api/v1
    description: Production
...
paths:
...
  /community/feed:
    get:
      tags: [community]
      summary: 获取信息流
      operationId: getFeed
      parameters:
        - name: tab
          in: query
          schema:
            type: string
            enum: [recommended, deep]
            default: recommended
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 信息流
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/FeedItem'
                          hasMore:
                            type: boolean
                          nextPage:
                            type: integer
  /community/post/{id}:
    get:
      tags: [community]
      summary: 获取帖子详情
      operationId: getPostDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 帖子详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FeedItem'
  /community/post/{id}/comments:
    get:
      tags: [community]
      summary: 获取评论列表
      operationId: getComments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 评论列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/Comment'
                          hasMore:
                            type: boolean
                          nextPage:
                            type: integer
    post:
      tags: [community]
      summary: 发表评论
      operationId: createComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  maxLength: 500
                parentId:
                  type: string
                  description: 回复的评论 ID（可选）
      responses:
        '200':
          description: 评论已创建
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Comment'
  /community/post:
    post:
      tags: [community]
      summary: 发布帖子
      operationId: createPost
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cardId:
                  type: string
                  description: 关联解读卡 ID（可选）
                shareText:
                  type: string
                  maxLength: 2000
                coverImageUrl:
                  type: string
      responses:
        '200':
          description: 帖子已发布
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                          status:
                            type: string
  /community/post/{id}/like:
    post:
      tags: [community]
      summary: 点赞
      operationId: likePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 点赞成功（幂等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /community/post/{id}/unlike:
    post:
      tags: [community]
      summary: 取消点赞
      operationId: unlikePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 取消成功（幂等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /community/post/{id}/favorite:
    post:
      tags: [community]
      summary: 收藏
      operationId: favoritePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 收藏成功（幂等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /community/post/{id}/unfavorite:
    post:
      tags: [community]
      summary: 取消收藏
      operationId: unfavoritePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 取消成功（幂等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /community/post/{id}/report:
    post:
      tags: [community]
      summary: 举报帖子
      operationId: reportPost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  enum: [porn, spam, abuse, other]
                detail:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: 举报已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /community/post/{id}/hide:
    post:
      tags: [community]
      summary: 隐藏帖子（减少推荐）
      operationId: hidePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 已隐藏
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /community/author/{authorId}:
    get:
      tags: [community]
      summary: 获取作者主页
      operationId: getAuthorProfile
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 作者信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthorProfile'
  /community/author/{authorId}/follow:
    post:
      tags: [community]
      summary: 关注作者
      operationId: followAuthor
      security:
        - bearerAuth: []
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 关注成功（幂等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /community/author/{authorId}/unfollow:
    post:
      tags: [community]
      summary: 取消关注
      operationId: unfollowAuthor
      security:
        - bearerAuth: []
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 取消成功（幂等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /community/author/{authorId}/block:
    post:
      tags: [community]
      summary: 屏蔽用户
      operationId: blockUser
      security:
        - bearerAuth: []
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 屏蔽成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /community/author/{authorId}/unblock:
    post:
      tags: [community]
      summary: 取消屏蔽
      operationId: unblockUser
      security:
        - bearerAuth: []
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 取消成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /community/search:
    get:
      tags: [community]
      summary: 搜索
      operationId: search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [all, post, user, activity]
            default: all
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - $ref: '#/components/schemas/PaginatedList'
  /community/feed/by-tag:
    get:
      tags: [community]
      summary: 按标签获取信息流
      operationId: getFeedByTag
      parameters:
        - name: tag
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 标签信息流
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - $ref: '#/components/schemas/PaginatedList'
  /community/tags/subscribe:
    post:
      tags: [community]
      summary: 订阅标签
      operationId: subscribeTag
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag]
              properties:
                tag:
                  type: string
      responses:
        '200':
          description: 订阅成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ─────────────── Notifications ───────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema:
        type: string
        format: uuid
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    FeedItem:
      type: object
      properties:
        id:
          type: string
        cardId:
          type: string
          nullable: true
        shareText:
          type: string
        coverImageUrl:
          type: string
          nullable: true
        authorId:
          type: string
        authorUsername:
          type: string
        authorHandle:
          type: string
          nullable: true
        authorAvatarUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          type: string
        metrics:
          type: object
          properties:
            likes:
              type: integer
            favorites:
              type: integer
            views:
              type: integer
            comments:
              type: integer
        viewerState:
          type: object
          properties:
            liked:
              type: boolean
            favorited:
              type: boolean
            followedAuthor:
              type: boolean
    Comment:
      type: object
      properties:
        id:
          type: string
        postId:
          type: string
        authorId:
          type: string
        authorUsername:
          type: string
        authorAvatarUrl:
          type: string
        parentId:
          type: string
          nullable: true
        text:
          type: string
        createdAt:
          type: string
          format: date-time
    AuthorProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        handle:
          type: string
        postCount:
          type: integer
        followerCount:
          type: integer
        followingCount:
          type: integer
        viewerState:
          type: object
          properties:
            followed:
              type: boolean
            blocked:
              type: boolean
...
    PaginatedList:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
            hasMore:
              type: boolean
            nextPage:
              type: integer